% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-france-official-emmergency.R
\name{getFranceOfficialEmmergency}
\alias{getFranceOfficialEmmergency}
\title{getFranceOfficialEmmergency}
\usage{
getFranceOfficialEmmergency(
 
    emergency_dep_url = "https://www.data.gouv.fr/fr/datasets/r/eceb9fb4-3ebc-4da3-828d-f5939712600a",
 
    emergency_dep_codebook_url = "https://www.data.gouv.fr/fr/datasets/r/2265d880-5260-4b34-8827-b3d76180032e",
 
    emergency_reg_url = "https://www.data.gouv.fr/fr/datasets/r/d2af5160-a21d-47b7-8f30-3c20dade63b1",
 
    emergency_reg_codebook_url = "https://www.data.gouv.fr/fr/datasets/r/02c99d99-96a0-4953-8e30-04fd660530f6",
 
    emergency_total_url = "https://www.data.gouv.fr/fr/datasets/r/219427ba-7e90-4eb1-9ac7-4de2e7e2112c",
 
    emergency_total_codebook_url = "https://www.data.gouv.fr/fr/datasets/r/0289e4fe-bb67-42b9-a320-cd22c98d82c3",
 
    age_categories_url = "https://www.data.gouv.fr/fr/datasets/r/cbed61a8-64e8-4c8c-8e97-65d3a31a99b6",
 
    regions_url = "https://data.opendatasoft.com/explore/dataset/georef-france-region@datanova/download/?format=csv&timezone=Europe/Berlin&lang=fr&use_labels_for_header=true&csv_separator=\%3B",
 
    region_departements_url = "https://www.data.gouv.fr/en/datasets/r/987227fb-dcb2-429e-96af-8979f97c9c84"
)
}
\value{
a list of two objects
}
\description{
A function to get France official emmergency data
}
\examples{
\dontrun{
 getFranceOfficialEmmergency()
}
}
