% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-france-official-hosp.R
\name{getFranceOfficialHosp}
\alias{getFranceOfficialHosp}
\title{getFranceOfficialHosp}
\usage{
getFranceOfficialHosp(
 
    hospital_dep_url = "https://www.data.gouv.fr/fr/datasets/r/63352e38-d353-4b54-bfd1-f1b3ee1cabd7",
 
    hospital_dep_codebook_url = "https://www.data.gouv.fr/fr/datasets/r/3f0f1885-25f4-4102-bbab-edec5a58e34a",
 
    france_region_departements_url = "https://www.data.gouv.fr/en/datasets/r/987227fb-dcb2-429e-96af-8979f97c9c84"
)
}
\value{


a list of two objects, cases_cum, cases_daily for France
}
\description{
A function to get France official hospitalization data
}
\examples{
\dontrun{
 getFranceOfficialHosp()
}
}
