% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-france-official-tests.R
\name{getFranceOfficialTests}
\alias{getFranceOfficialTests}
\title{getFranceOfficialTests}
\usage{
getFranceOfficialTests(
 
    tests_dep_url = "https://www.data.gouv.fr/fr/datasets/r/b4ea7b4b-b7d1-4885-a099-71852291ff20",
 
    tests_dep_codebook_url = "https://www.data.gouv.fr/fr/datasets/r/971c5cbd-cd80-4492-b2b3-c3deff8c1f5e",
 
    regions_url = "https://data.opendatasoft.com/explore/dataset/contours-geographiques-tres-simplifies-des-regions-2019@ofgl-opendatamef/download/?format=csv&timezone=Europe/Berlin&use_labels_for_header=true&csv_separator=\%3B",
 
    region_departements_url = "https://www.data.gouv.fr/en/datasets/r/987227fb-dcb2-429e-96af-8979f97c9c84"
)
}
\value{
a list of two objects
}
\description{
A function to get France official Tests data
}
\examples{
\dontrun{
 getFranceOfficialEmmergency()
}
}
