% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-jchu-cases.R
\name{getJHUcases}
\alias{getJHUcases}
\title{getJHUcases}
\usage{
getJHUcases(
 
    jhu_data_cases_url = "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv",
 
    jhu_data_deaths_url = "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_global.csv"
)
}
\value{


a list of three objects, cases_cum, cases_daily, deaths_cum per
countries
}
\description{
A function to get covid-19 cases from github
}
\examples{
\dontrun{
 getJHUcases()
}
}
