% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-opencovidfr.R
\name{getOpenCovidFr}
\alias{getOpenCovidFr}
\title{getOpenCovidFr}
\usage{
getOpenCovidFr(
 
    france_com_url = "https://www.data.gouv.fr/fr/datasets/r/0b66ca39-1623-4d9c-83ad-5434b7f9e2a4"
)
}
\value{


a list of two objects, cases_cum, cases_daily for France
}
\description{
A function to get covid-19 cases from opencovid france
}
\examples{
\dontrun{
 getOpenCovidFr()
}
}
