% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_plot.R
\name{lambda_plot}
\alias{lambda_plot}
\title{lambda_plot}
\usage{
lambda_plot(my_country, mydata, my_date, si_mean = 5, si_sd = 3.4)
}
\arguments{
\item{my_country}{Name of the country as chr}

\item{mydata}{daily incident cases}

\item{my_date}{starting date for the plot}
}
\value{
a plot
}
\description{
a function to plot lambda function showing epidemics intensity
}
