% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEpicurve.R
\name{makeEpicurve}
\alias{makeEpicurve}
\title{makeEpicurve}
\usage{
makeEpicurve(my_data = NULL, my_ncol = 3)
}
\arguments{
\item{my_data}{the cases data, daily}

\item{my_ncol}{the number of columns to be dsiplayed}
}
\value{
a ggplot object
}
\description{
a function to plot epidemic curves for a set of countries
}
