% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ft-log-trends.R
\name{plotFTlogTrendCases}
\alias{plotFTlogTrendCases}
\title{plotFTlogTrendCases}
\usage{
plotFTlogTrendCases(jhu_data = NULL, start_case = 100)
}
\value{
a list containing the data and the ggplot object
}
\description{
A function to get covid-19 cases from github
}
\details{
Total Number of Covid-19 cases per country, log scale, from 100+ plus
Original code from https://twitter.com/jburnmurdoch
Stories, stats & scatterplots for FinancialTimes
john.burn-murdoch(at)ft.com
}
\examples{
\dontrun{
 plotFTlogTrendCases(jhu_data = mydata)
}
}
\seealso{
\url{https://gist.githubusercontent.com/johnburnmurdoch/34bd7470dca92e470fd5f12a488923ce/raw/c7a66c8a718299ecfdb9b7c922daca5d33eb8aa0/coronavirus_cases_trajectories.R}
}
