% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSmallMultipleCovid.R
\name{plotSmallMultipleCovid}
\alias{plotSmallMultipleCovid}
\title{plotSmallMultipleCovid}
\usage{
plotSmallMultipleCovid(
  my_data = NULL,
  myvariable = "death_cum",
  colfeature = "region_name",
  facetfeature = "dep_name",
  facetscales = "free_y",
  ncol = 10,
  mytitle_base = "France per departments : "
)
}
\arguments{
\item{mytitle_base}{}
}
\value{

}
\description{
plotSmallMultipleCovid
}
